text: switch to an atlas based implementation
This new implementation is motivated by a two main reasons:
  - the texture can now be shared between all the Text instances
  - live changing becomes possible without reconstructing a texture

The ref difference is because of precision inaccuracies in the current
implementation:

In test-text-align_cc, if we remove the font_scale and set the padding
to 0, we will observe a 1280x800 export with a vertically centered text
area of 1280x392. In this situation, the y offset happens to be 196
instead of 204 for a properly vertically centered text
(204+392+204=800). This is due to the rounding in the padh / 2
operation: the texture is much smaller that the export (padh=25), so
padh/2 will lead to either 12 (194 in the export) or 13 (212 in the
export) if we were using (padh+1)/2. This is the root cause of the
differences in the references, and it can not be adressed easily in the
current implementation. The new implementation is not affected by this
precision issue.