compute: add ComputeProgram.workgroup_size parameter
The name is chosen to match Khronos nomenclature used in the
limits/capabilities API.

The parameter could have been set in the Compute node (along with
workgroup_count) instead of ComputeProgram. But ComputeProgram is much
more likely to be shared than Compute, so that would imply having
the usually identical workgroup sizes duplicated at every Compute. Which
means that if the shader code needs to be adjusted to work with a
different local size layout, all the Compute would be required to be
adjusted. This duplication, and the risk of de-synchronization implied
between the shader code and the user nodes seemed fragile, so it was
decided to have it in ComputeProgram instead, where it is very close to
the shader code.