block: move type field to pgcraft_block structure
This commits partially reverts the changes introduced by
fe61ad4acc960d1f517d6db49f6b80187f2c43db. The main difference is that we
are not overriding the type of the shared block anymore. This allows the
underlying buffer to be bound as a uniform buffer or a storage buffer
depending on the pipeline needs (buffer is writable or not) and the
buffer constrains (layout, size).